+++ changelog.md
@@ -26,3 +26,5 @@
 - changed `text_below_button` to `text_2` in `price-card`
 - `tabbed_content_section` `entries` field changed to max 3
 - removed `two-col-image-text-section` favor of `two-columns-section`
+- new `faq-section`
+- new `testimonials-section`